// backend/src/utils/enums/permissions.ts
export enum Permissions {
    MANAGE_USERS = "MANAGE_USERS",
    VIEW_DASHBOARD = "VIEW_DASHBOARD",
    MANAGE_STOCK = "MANAGE_STOCK",
    MANAGE_PURCHASES = "MANAGE_PURCHASES",
    MANAGE_RECIPES = "MANAGE_RECIPES",
    MANAGE_SALES = "MANAGE_SALES",
    AUDIT_LOGS = "AUDIT_LOGS",
    READ_ONLY = "READ_ONLY"
}